// components/block/regulations/types.ts
export interface RegularPost {
    id: string;
    title: string;
    meta: string;
    description: string;
    url: string;
    featured: string;
    tech: string[];
}

export interface SpecialPost {
    id: string;
    title: string;
    meta: string;
    description: string;
    url: string;
    featured: string;
    tech: string[];
}